修复issue#2:重启报错问题

+++ myblog/setting.py
@@ -22,7 +22,8 @@ class BaseConfig:
 
     for i in os.listdir(USER_DEFAULT_DIR):
         src = os.path.join(USER_DEFAULT_DIR, i)
-        shutil.move(src, USERDATA_DIR)
+        if not os.path.exists(src):
+            shutil.move(src, USERDATA_DIR)
 
 
 class DevelopmentConfig(BaseConfig):
